Fix building built-in hashes on Windows

+++ trantor/utils/crypto/sha1.h
@@ -9,16 +9,18 @@ Last modified by Martin Chang for the Trantor project
 
 #pragma once
 
+#include <stdint.h>
+
 typedef struct
 {
-    u_int32_t state[5];
-    u_int32_t count[2];
+    uint32_t state[5];
+    size_t count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 
-void TrantorSHA1Transform(u_int32_t state[5], const unsigned char buffer[64]);
+void TrantorSHA1Transform(uint32_t state[5], const unsigned char buffer[64]);
 void TrantorSHA1Init(SHA1_CTX* context);
 void TrantorSHA1Update(SHA1_CTX* context,
                        const unsigned char* data,
-                       u_int32_t len);
+                       size_t len);
 void TrantorSHA1Final(unsigned char digest[20], SHA1_CTX* context);
